6300064 hsfs_spec.h has developed a southern drawl. 6338131 IDE_SUA_LEN() macro can cause integer underflow 6339513 strlcat() isn't always appropriate ... 6343199 hsfs mount error exit part still doesn't clean up correctly

+++ usr/src/uts/common/sys/fs/hsfs_isospec.h
@@ -2,9 +2,8 @@
  * CDDL HEADER START
  *
  * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
+ * Common Development and Distribution License (the "License").
+ * You may not use this file except in compliance with the License.
  *
  * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
  * or http://www.opensolaris.org/os/licensing.
@@ -20,7 +19,7 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2004 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
@@ -218,8 +217,8 @@ enum iso_voldesc_type {
 #define	IDE_NAME_LEN(x)		*(IDE_name_len(x))
 #define	IDE_NAME(x)		IDE_name(x)
 #define	IDE_PAD_LEN(x)		((IDE_NAME_LEN(x) % 2) ? 0 : 1)
-#define	IDE_SUA_LEN(x)		(IDE_DIR_LEN(x) - IDE_FDESIZE - \
-					IDE_NAME_LEN(x) - IDE_PAD_LEN(x))
+#define	IDE_SUA_LEN(x)		((int)(IDE_DIR_LEN(x)) - (int)(IDE_FDESIZE) - \
+				(int)(IDE_NAME_LEN(x)) - (int)(IDE_PAD_LEN(x)))
 
 /* mask bits for IDE_FLAGS */
 #define	IDE_EXISTENCE		0x01	/* zero if file exists */
